refactor: nest style based on dimension
