+++ src/routes/+page.svelte
@@ -3,7 +3,7 @@
 
 	import Button from '$lib/components/ui/button.svelte'
 	import Checkbox from '$lib/components/ui/checkbox.svelte'
-	import Input from '$lib/components/input.svelte'
+	import Input from '$lib/components/ui/input.svelte'
 	import Radio from '$lib/components/ui/radio.svelte'
 	import Typography from '$lib/components/ui/typography.svelte'
 	import Switch from '$lib/components/ui/switch.svelte'
@@ -306,9 +306,9 @@
 	</div>
 </section>
 <Input placeholder="Input label">This is some helper text.</Input>
-<Input placeholder="Input label" layout="large">This is some helper text.</Input>
-<Input placeholder="Input label" layout="compact" active>This is some helper text.</Input>
-<Input placeholder="Input label" layout="small">This is some helper text.</Input>
+<Input placeholder="Input label" dimension="large">This is some helper text.</Input>
+<Input placeholder="Input label" dimension="compact" active>This is some helper text.</Input>
+<Input placeholder="Input label" dimension="small">This is some helper text.</Input>
 
 <style>
 	.row {
